Updated to 1.6.1, performed cleanup and javadoc-ing

+++ src/main/java/ch/spacebase/mcprotocol/standard/packet/PacketHealthUpdate.java
@@ -10,14 +10,14 @@ import ch.spacebase.mcprotocol.packet.Packet;
 
 public class PacketHealthUpdate extends Packet {
 
-	public short health;
+	public float health;
 	public short food;
 	public float saturation;
 
 	public PacketHealthUpdate() {
 	}
 
-	public PacketHealthUpdate(short health, short food, float saturation) {
+	public PacketHealthUpdate(float health, short food, float saturation) {
 		this.health = health;
 		this.food = food;
 		this.saturation = saturation;
@@ -25,14 +25,14 @@ public class PacketHealthUpdate extends Packet {
 
 	@Override
 	public void read(DataInputStream in) throws IOException {
-		this.health = in.readShort();
+		this.health = in.readFloat();
 		this.food = in.readShort();
 		this.saturation = in.readFloat();
 	}
 
 	@Override
 	public void write(DataOutputStream out) throws IOException {
-		out.writeShort(this.health);
+		out.writeFloat(this.health);
 		out.writeShort(this.food);
 		out.writeFloat(saturation);
 	}
