+++ src/main/java/ch/spacebase/mcprotocol/standard/packet/PacketIncrementStatistic.java
@@ -11,12 +11,12 @@ import ch.spacebase.mcprotocol.packet.Packet;
 public class PacketIncrementStatistic extends Packet {
 
 	public int statistic;
-	public byte amount;
+	public int amount;
 
 	public PacketIncrementStatistic() {
 	}
 
-	public PacketIncrementStatistic(int statistic, byte amount) {
+	public PacketIncrementStatistic(int statistic, int amount) {
 		this.statistic = statistic;
 		this.amount = amount;
 	}
@@ -24,13 +24,13 @@ public class PacketIncrementStatistic extends Packet {
 	@Override
 	public void read(DataInputStream in) throws IOException {
 		this.statistic = in.readInt();
-		this.amount = in.readByte();
+		this.amount = in.readInt();
 	}
 
 	@Override
 	public void write(DataOutputStream out) throws IOException {
 		out.writeInt(this.statistic);
-		out.writeByte(this.amount);
+		out.writeInt(this.amount);
 	}
 
 	@Override
